+++ Source/filter.py
@@ -15,13 +15,13 @@ import searchOptions
 def _filterList_(name,valueList):
     defaultReturn = {}
     if len(valueList) == 1:
-        defaultReturn = {"field": name, "operator": "is", "value": valueList[0]}
+        defaultReturn = {"operator": "is", "field": name, "value": valueList[0]}
     elif len(valueList) > 1:
         for x in valueList:
             if defaultReturn == {}:
                 defaultReturn = []
             defaultReturn.append(x)
-        defaultReturn = {"field": name, "operator": "contains", "value": defaultReturn}
+        defaultReturn = {"operator": "contains", "field": name, "value": defaultReturn}
     return defaultReturn
 
 # Filter function
@@ -29,14 +29,11 @@ def _filterList_(name,valueList):
 def filter(options: searchOptions):
     genreFilter = _filterList_('genre', options.getGenre())
     tagFilter = _filterList_('tag', options.getTag())
-    castFilter = _filterList_('cast', options.getCast())
+    castFilter = _filterList_('actor', options.getCast())
     directorFilter = _filterList_('director', options.getDirector())
-    #includeFilter = _filterList_(options.getInclude())
-    #excludeFilter = _filterList_(options.getExclude())
-    #lengthFilter = _filterList_(options.getLength())
     yearFilter = _filterList_('year', options.getYear())
     studioFilter = _filterList_('studio', options.getStudio())
-    ratingFilter = _filterList_('rating', options.getRating())
+    ratingFilter = _filterList_('mpaarating', options.getRating())
 
     filterList = []
     if genreFilter != {}:
@@ -47,20 +44,14 @@ def filter(options: searchOptions):
         filterList.append(castFilter)
     if directorFilter != {}:
         filterList.append(directorFilter)
-    '''
-    if includeFilter != {}:
-        filterList.append(includeFilter)
-    if excludeFilter != {}:
-        filterList.append(excludeFilter)
-    if lengthFilter != {}:
-        filterList.append(lengthFilter)
-    '''
-    if yearFilter != {}:
-        filterList.append(yearFilter)
+    '''if yearFilter != {}:
+        filterList.append(yearFilter)'''
     if studioFilter != {}:
         filterList.append(studioFilter)
     if ratingFilter != {}:
         filterList.append(ratingFilter)
+
+    print (filterList)
 
     # if movie type
     # function call to get movie list
@@ -68,33 +59,48 @@ def filter(options: searchOptions):
     # function call to get episode list
     # combine show lists
     # trigger playback functions
-    videoList = []
+    mediaTypes = options.getMediaType()
+    movieList = []
+    episodeList = []
+    for x in mediaTypes:
+        if(x == 'movie'):
+            command = {
+                "jsonrpc": "2.0",
+                "method": "VideoLibrary.GetMovies",
+                "params": {
+                    "filter": { "or" : filterList }, # get_genre_filer() = {"field": "genre", "operator": "is", "value": genre}
+                    "properties": ["art", "thumbnail", "playcount", "file", "runtime", "genre", "tag", "cast", "director", "year", "studio", "mpaa", "rating", "userrating"],
+                    "sort": {"order": "ascending", "method": "label"}
+                },
+                "id": "library"}
+            print(command)
+            movieList = json.loads(xbmc.executeJSONRPC(json.dumps(command)))
+        elif(x == 'episode'):
+            command = {
+                "jsonrpc": "2.0",
+                "method": "VideoLibrary.GetEpisodes",
+                "params": {
+                    "filter": { "or" : filterList }, # get_genre_filer() = {"field": "genre", "operator": "is", "value": genre}
+                    "properties": ["art", "rating", "thumbnail", "playcount", "file", "genre"],
+                    "sort": {"order": "ascending", "method": "label"}
+                },
+                "id": "library"}
+            print(command)
+            episodeList = json.loads(xbmc.executeJSONRPC(json.dumps(command)))
+
+    #videoList = episodeList + movieList
+    print(episodeList)
+    print(movieList)
+
+    '''videoList = filterIncExc(videoList, options.getInclude(), options.getExclude())
+    videoList = filterLength(videoList, options.getInclude(), options.getExclude())
+
     if options.getPBFunction() == 1:
         playOne(videoList)
     elif options.getPBFunction() == 2:
         showList(videoList)
     elif options.getPBFunction() == 3:
-        loopPlay(videoList)
-
-    command = {
-        "jsonrpc": "2.0",
-        "method": "VideoLibrary.GetEpisodes",
-        "params": {
-            "filter": { "or" : filterList }, # get_genre_filer() = {"field": "genre", "operator": "is", "value": genre}
-            "properties": ["art", "rating", "thumbnail", "playcount", "file"],
-            "sort": {"order": "ascending", "method": "label"}
-        },
-        "id": "library"}
-    
-    response = xbmc.executeJSONRPC(json.dumps(command))
-    response_obj = json.loads(response)
-    if "result" in response_obj:
-        items = response_obj["result"]["episodes"]
-        items.extend(response_obj["result"]["movies"])
-        return items
-    else:
-        print("no item found")
-        return {}
+        loopPlay(videoList)'''
 
 
 def unitTest(options: searchOptions):
@@ -105,3 +111,5 @@ def unitTest(options: searchOptions):
     print(_filterList_('year', options.getYear()))
     print(_filterList_('studio', options.getStudio()))
     print(_filterList_('rating', options.getRating()))
+
+    filter(options)